requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope
asttokens
autograd==1.6.2
backcall
boltons==23.0.0
brotlipy
certifi==2022.12.7
cffi
charset-normalizer
comm
contourpy
control
cryptography
cycler
debugpy
decorator
distinctipy==1.3.4
distro==1.9.0
executing
filelock
fonttools
future
gmpy2
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna
importlib-metadata
ipykernel
ipython
ipywidgets==8.0.6
jedi
Jinja2==3.1.2
jupyter_client
jupyter_core
jupyterlab-widgets==3.0.7
kiwisolver
MarkupSafe==2.1.2
matplotlib
matplotlib-inline
mpmath==1.2.1
munkres==1.1.4
nest-asyncio
networkx
numpy
openai==1.12.0
packaging
parso
pexpect
pickleshare
Pillow
platformdirs
pooch
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
pydantic==2.6.1
pydantic_core==2.16.2
Pygments
pyOpenSSL
pyparsing
PySocks
python-dateutil
PyYAML
pyzmq
requests
scipy
sdeint==0.3.0
six
slycot
sniffio==1.3.0
stack-data
sympy
torch==2.0.1
torchaudio==2.0.1
torchdiffeq
torchsde==0.2.5
torchvision==0.15.1
tornado
tqdm==4.65.0
traitlets
trampoline==0.1.2
typing_extensions==4.9.0
urllib3
wcwidth
widgetsnbextension==4.0.7
zipp
